Refactor addSystemModalReducer
Moved addSystemModalReducer into it's own file. Will do this with other reducers if we are ok with this setup.
