+++ src/SmartComponents/AddSystemModal/AddSystemModal.js
@@ -35,7 +35,6 @@ class AddSystemModal extends Component {
     }
 
     render() {
-
         return (
             <React.Fragment>
                 <Modal
