+++ src/SmartComponents/modules/addSystemModalReducer.js
@@ -1,0 +1,24 @@
+import types from './types';
+
+const initialState = {
+    addSystemModalOpened: false
+};
+
+function addSystemModalReducer(state = initialState, action) {
+    switch (action.type) {
+        case `${types.OPEN_ADD_SYSTEM_MODAL}`:
+            return {
+                ...state,
+                addSystemModalOpened: !state.addSystemModalOpened
+            };
+
+        default:
+            return {
+                ...state
+            };
+    }
+}
+
+export default {
+    addSystemModalReducer
+};
