+++ src/SmartComponents/modules/index.js
@@ -1,3 +1,4 @@
 export { default as compareActions } from './actions';
 export { default as compareTypes } from './types';
 export { default as reducers } from './reducers';
+export { default as addSystemModalReducer } from './addSystemModalReducer';
